exitOnError when linking forge dependencies
