+++ packages/api/core/src/api/init.ts
@@ -12,6 +12,7 @@ import { readRawPackageJson } from '../util/read-package-json';
 import { findTemplate } from './init-scripts/find-template';
 import { initDirectory } from './init-scripts/init-directory';
 import { initGit } from './init-scripts/init-git';
+import { initLink } from './init-scripts/init-link';
 import { initNPM } from './init-scripts/init-npm';
 
 const d = debug('electron-forge:init');
@@ -108,6 +109,7 @@ export default async ({ dir = process.cwd(), interactive = false, copyCIFiles = 
                   }
                   return await installDepList(dir, templateModule.dependencies || [], DepType.PROD, DepVersionRestriction.RANGE);
                 },
+                exitOnError: false,
               },
               {
                 title: 'Installing development dependencies',
@@ -118,17 +120,32 @@ export default async ({ dir = process.cwd(), interactive = false, copyCIFiles = 
                   }
                   await installDepList(dir, templateModule.devDependencies || [], DepType.DEV);
                 },
+                exitOnError: false,
               },
               {
                 title: 'Finalizing dependencies',
                 task: async (_, task) => {
-                  await initNPM(dir, task);
+                  return task.newListr([
+                    {
+                      title: 'Installing common dependencies',
+                      task: async (_, task) => {
+                        await initNPM(dir, task);
+                      },
+                      exitOnError: false,
+                    },
+                    {
+                      title: 'Maybe linking forge dependencies',
+                      task: async (_, task) => {
+                        await initLink(dir, task);
+                      },
+                      exitOnError: true,
+                    },
+                  ]);
                 },
               },
             ],
             {
               concurrent: false,
-              exitOnError: false,
             }
           );
         },
